update cat-data-service API
