+++ src/cat-data-service/src/service/v1/event/mod.rs
@@ -13,28 +13,22 @@ use std::sync::Arc;
 use super::LimitOffset;
 
 mod objective;
-mod proposal;
-mod review;
 
 pub fn event(state: Arc<State>) -> Router {
     let objective = objective::objective(state.clone());
-    let proposal = proposal::proposal(state.clone());
-    let review = review::review(state.clone());
 
     Router::new()
         .nest(
-            "/event",
+            "/event/:event",
             Router::new()
                 .route(
-                    "/:event",
+                    "/",
                     get({
                         let state = state.clone();
                         move |path| async { handle_result(event_exec(path, state).await).await }
                     }),
                 )
-                .merge(objective)
-                .merge(proposal)
-                .merge(review),
+                .merge(objective),
         )
         .route(
             "/events",
