+++ src/cat-data-service/src/service/v1/event/objective/review_type.rs
@@ -7,50 +7,16 @@ use axum::{
     routing::get,
     Router,
 };
-use event_db::types::event::{
-    objective::ObjectiveId,
-    proposal::ProposalId,
-    review::{AdvisorReview, ReviewType},
-    EventId,
-};
+use event_db::types::event::{objective::ObjectiveId, review::ReviewType, EventId};
 use std::sync::Arc;
 
-pub fn review(state: Arc<State>) -> Router {
-    Router::new()
-        .route(
-            "/:event/:objective/:proposal/reviews",
-            get({
-                let state = state.clone();
-                move |path, query| async {
-                    handle_result(reviews_exec(path, query, state).await).await
-                }
-            }),
-        )
-        .route(
-            "/:event/:objective/review_types",
-            get(move |path, query| async {
-                handle_result(review_types_exec(path, query, state).await).await
-            }),
-        )
-}
-
-async fn reviews_exec(
-    Path((event, objective, proposal)): Path<(EventId, ObjectiveId, ProposalId)>,
-    lim_ofs: Query<LimitOffset>,
-    state: Arc<State>,
-) -> Result<Vec<AdvisorReview>, Error> {
-    tracing::debug!(
-        "reviews_query, event:{0} objective: {1}, proposal: {2}",
-        event.0,
-        objective.0,
-        proposal.0,
-    );
-
-    let reviews = state
-        .event_db
-        .get_reviews(event, objective, proposal, lim_ofs.limit, lim_ofs.offset)
-        .await?;
-    Ok(reviews)
+pub fn review_type(state: Arc<State>) -> Router {
+    Router::new().route(
+        "/review_types",
+        get(move |path, query| async {
+            handle_result(review_types_exec(path, query, state).await).await
+        }),
+    )
 }
 
 async fn review_types_exec(
@@ -86,146 +52,8 @@ mod tests {
         body::{Body, HttpBody},
         http::{Request, StatusCode},
     };
-    use event_db::types::event::review::Rating;
     use serde_json::json;
     use tower::ServiceExt;
-
-    #[tokio::test]
-    async fn reviews_test() {
-        let state = Arc::new(State::new(None).await.unwrap());
-        let app = app(state);
-
-        let request = Request::builder()
-            .uri(format!("/api/v1/event/{0}/{1}/{2}/reviews", 1, 1, 1))
-            .body(Body::empty())
-            .unwrap();
-        let response = app.clone().oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-
-        assert_eq!(
-            serde_json::to_string(&vec![
-                AdvisorReview {
-                    assessor: "assessor 1".to_string(),
-                    ratings: vec![
-                        Rating {
-                            review_type: 1,
-                            score: 10,
-                            note: Some("note 1".to_string()),
-                        },
-                        Rating {
-                            review_type: 2,
-                            score: 15,
-                            note: Some("note 2".to_string()),
-                        },
-                        Rating {
-                            review_type: 5,
-                            score: 20,
-                            note: Some("note 3".to_string()),
-                        }
-                    ],
-                },
-                AdvisorReview {
-                    assessor: "assessor 2".to_string(),
-                    ratings: vec![],
-                },
-                AdvisorReview {
-                    assessor: "assessor 3".to_string(),
-                    ratings: vec![],
-                },
-            ])
-            .unwrap(),
-            String::from_utf8(response.into_body().data().await.unwrap().unwrap().to_vec())
-                .unwrap()
-        );
-
-        let request = Request::builder()
-            .uri(format!(
-                "/api/v1/event/{0}/{1}/{2}/reviews?lim={3}",
-                1, 1, 1, 2
-            ))
-            .body(Body::empty())
-            .unwrap();
-        let response = app.clone().oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-
-        assert_eq!(
-            serde_json::to_string(&vec![
-                AdvisorReview {
-                    assessor: "assessor 1".to_string(),
-                    ratings: vec![
-                        Rating {
-                            review_type: 1,
-                            score: 10,
-                            note: Some("note 1".to_string()),
-                        },
-                        Rating {
-                            review_type: 2,
-                            score: 15,
-                            note: Some("note 2".to_string()),
-                        },
-                        Rating {
-                            review_type: 5,
-                            score: 20,
-                            note: Some("note 3".to_string()),
-                        }
-                    ],
-                },
-                AdvisorReview {
-                    assessor: "assessor 2".to_string(),
-                    ratings: vec![],
-                },
-            ])
-            .unwrap(),
-            String::from_utf8(response.into_body().data().await.unwrap().unwrap().to_vec())
-                .unwrap()
-        );
-
-        let request = Request::builder()
-            .uri(format!(
-                "/api/v1/event/{0}/{1}/{2}/reviews?ofs={3}",
-                1, 1, 1, 1
-            ))
-            .body(Body::empty())
-            .unwrap();
-        let response = app.clone().oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-
-        assert_eq!(
-            serde_json::to_string(&vec![
-                AdvisorReview {
-                    assessor: "assessor 2".to_string(),
-                    ratings: vec![],
-                },
-                AdvisorReview {
-                    assessor: "assessor 3".to_string(),
-                    ratings: vec![],
-                },
-            ])
-            .unwrap(),
-            String::from_utf8(response.into_body().data().await.unwrap().unwrap().to_vec())
-                .unwrap()
-        );
-
-        let request = Request::builder()
-            .uri(format!(
-                "/api/v1/event/{0}/{1}/{2}/reviews?lim={3}&ofs={4}",
-                1, 1, 1, 1, 1
-            ))
-            .body(Body::empty())
-            .unwrap();
-        let response = app.clone().oneshot(request).await.unwrap();
-        assert_eq!(response.status(), StatusCode::OK);
-
-        assert_eq!(
-            serde_json::to_string(&vec![AdvisorReview {
-                assessor: "assessor 2".to_string(),
-                ratings: vec![],
-            },])
-            .unwrap(),
-            String::from_utf8(response.into_body().data().await.unwrap().unwrap().to_vec())
-                .unwrap()
-        );
-    }
 
     #[tokio::test]
     async fn review_types_test() {
@@ -233,7 +61,10 @@ mod tests {
         let app = app(state);
 
         let request = Request::builder()
-            .uri(format!("/api/v1/event/{0}/{1}/review_types", 1, 1))
+            .uri(format!(
+                "/api/v1/event/{0}/objective/{1}/review_types",
+                1, 1
+            ))
             .body(Body::empty())
             .unwrap();
         let response = app.clone().oneshot(request).await.unwrap();
@@ -287,7 +118,7 @@ mod tests {
 
         let request = Request::builder()
             .uri(format!(
-                "/api/v1/event/{0}/{1}/review_types?lim={2}",
+                "/api/v1/event/{0}/objective/{1}/review_types?lim={2}",
                 1, 1, 2
             ))
             .body(Body::empty())
@@ -325,7 +156,7 @@ mod tests {
 
         let request = Request::builder()
             .uri(format!(
-                "/api/v1/event/{0}/{1}/review_types?ofs={2}",
+                "/api/v1/event/{0}/objective/{1}/review_types?ofs={2}",
                 1, 1, 1
             ))
             .body(Body::empty())
@@ -371,7 +202,7 @@ mod tests {
 
         let request = Request::builder()
             .uri(format!(
-                "/api/v1/event/{0}/{1}/review_types?lim={2}&ofs={3}",
+                "/api/v1/event/{0}/objective/{1}/review_types?lim={2}&ofs={3}",
                 1, 1, 1, 1
             ))
             .body(Body::empty())
